+++ packages/i18n/tsup.config.ts
@@ -15,21 +15,6 @@ const config: Options = {
   treeshake: true,
   skipNodeModulesBundle: true,
   external: ['node_modules'],
-  plugins: [
-    {
-      // https://github.com/egoist/tsup/issues/953
-      // ensuring that all local imports in `.js` files import from `.js` files.
-      name: 'fix-esm-js-imports',
-      renderChunk(code) {
-        if (this.format === 'esm') {
-          const regexEsm = /from(?<space>[\s]*)(?<quote>['"])(?<import>\.[^'"]+)['"]/g
-          return {
-            code: code.replace(regexEsm, 'from$<space>$<quote>$<import>.js$<quote>'),
-          }
-        }
-      },
-    },
-  ],
 }
 
 export default defineConfig([config])